% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-classes.R
\name{AnalysisResult}
\alias{AnalysisResult}
\title{AnalysisResult class}
\usage{
AnalysisResult(
  summary = NULL,
  method = "describe",
  params = list(),
  source = NULL
)
}
\arguments{
\item{summary}{A \code{data.frame} of computed statistics.}

\item{method}{Analysis method. One of \code{"describe"} or \code{"correlate"}.
Default is \code{"describe"}.}

\item{params}{A list of analysis parameters. Default is \code{list()}.}

\item{source}{Optional reference to the source \link{Dataset}. Default is \code{NULL}.}
}
\value{
An AnalysisResult object.
}
\description{
Container for the output of \code{\link[=analyse]{analyse()}}.
}
