% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Application UI}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Shiny request object (required by \code{shinyApp}).}
}
\value{
A \code{bslib} page definition.
}
\description{
Application UI
}
\keyword{internal}
