% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_server}
\alias{mod_analysis_server}
\title{Analysis module server}
\usage{
mod_analysis_server(id, dataset, result)
}
\arguments{
\item{id}{Module id.}

\item{dataset}{A \code{reactiveVal} holding the input \link{Dataset}.}

\item{result}{A \code{reactiveVal} to write the \link{AnalysisResult} into.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Analysis module server
}
\keyword{internal}
