% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_ui}
\alias{mod_analysis_ui}
\title{Analysis module UI}
\usage{
mod_analysis_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A UI definition.
}
\description{
Analysis module UI
}
\keyword{internal}
