% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_input_server}
\alias{mod_input_server}
\title{Input module server}
\usage{
mod_input_server(id, dataset)
}
\arguments{
\item{id}{Module id.}

\item{dataset}{A \code{reactiveVal} to write the loaded \link{Dataset} into.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Input module server
}
\keyword{internal}
