% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report module server}
\usage{
mod_report_server(id, result)
}
\arguments{
\item{id}{Module id.}

\item{result}{A \code{reactiveVal} holding the \link{AnalysisResult}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Report module server
}
\keyword{internal}
