% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Report module UI}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A UI definition.
}
\description{
Report module UI
}
\keyword{internal}
