% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-classes.R
\name{print.Dataset}
\alias{print.Dataset}
\title{Print a Dataset object}
\usage{
\method{print}{Dataset}(x, ...)
}
\arguments{
\item{x}{A Dataset object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Dataset object
}
\keyword{internal}
