% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-classes.R
\name{print.Project}
\alias{print.Project}
\title{Print a Project object}
\usage{
\method{print}{Project}(x, ...)
}
\arguments{
\item{x}{A Project object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Project object
}
\keyword{internal}
