% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-classes.R
\name{Project}
\alias{Project}
\title{Project class}
\usage{
Project(data = NULL, results = list(), name = "", description = "")
}
\arguments{
\item{data}{A \link{Dataset} object, or \code{NULL}.}

\item{results}{A list of \link{AnalysisResult} objects. Default is \code{list()}.}

\item{name}{Project name. Default is \code{""}.}

\item{description}{Project description. Default is \code{""}.}
}
\value{
A Project object.
}
\description{
A container for a complete analysis: an input \link{Dataset} plus a list of
\link{AnalysisResult} objects.
}
